Add game time remaining variable
