+++ src/variables/errors.ts
