+++ src/variables/index.ts
@@ -1,7 +1,8 @@
 import { TriviaGame } from '../globals';
 import { arrayJoinWith } from './array-join-with';
-import { mageTriviaError, mageTriviaErrorFull } from './mage-trivia-error';
-import { mageTriviaAnswers, mageTriviaCorrectAnswers, mageTriviaGameInProgress, mageTriviaGameResultsRaw, mageTriviaPossibleAnswers, mageTriviaQuestion, mageTriviaQuestionAndAnswersRaw, mageTriviaWinners, mageTriviaWinnersWithPoints } from './unorganized';
+import { mageTriviaError, mageTriviaErrorFull } from './errors';
+import { mageTriviaGameInProgress, mageTriviaGameTimeRemaining } from './game-status';
+import { mageTriviaAnswers, mageTriviaCorrectAnswers, mageTriviaGameResultsRaw, mageTriviaPossibleAnswers, mageTriviaQuestion, mageTriviaQuestionAndAnswersRaw, mageTriviaWinners, mageTriviaWinnersWithPoints } from './unorganized';
 import { mageTriviaAnswerAccepted, mageTriviaAnswerRejected, mageTriviaAnswerRejectedRaw } from './user-answer';
 
 export function registerReplaceVariables(triviaGame: TriviaGame): void {
@@ -12,6 +13,7 @@ export function registerReplaceVariables(triviaGame: TriviaGame): void {
     triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaError);
     triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaErrorFull);
     triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaGameInProgress);
+    triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaGameTimeRemaining);
     triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaQuestionAndAnswersRaw);
     triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaQuestion);
     triviaGame.getFirebotManager().registerReplaceVariable(mageTriviaAnswers);
